Rename `Constants.h` to `MemConstants.h`
